Always copy IPA images in bifrost container
When we build the IPA images we must always copy them in bifrost otherwise the will not copy if they were already built in the past. We meet this case when redeploying the seed, the images are not present in /httpboot/.
So we propose to systematically copy images.

Update seed-ipa-build.yml

update seed-ipa-build.yml file