Add empty kolla inventory group when group is specified but unmapped
In some cases we may have a kolla group defined in
'kolla_overcloud_inventory_top_level_group_map' which has no kayobe groups
mapped to it. This causes no entry to be added to the inventory for the group.
If the group is referenced elsewhere in the inventory, then Ansible will fail
to parse the inventory file.

The specific case that failed was when setting
'kolla_overcloud_inventory_storage_groups' to an empty list.

This change ensures that we add an empty definition for the group in the kolla
inventory.

Story: 2002105
Task: 19783