Update python dependencies in kolla-ansible virtualenv
If upgrading the kayobe control host environment via kayobe control host
upgrade, kolla-ansible will be upgraded in its virtual environment if the
specified source repo and version are different than those installed. However,
other dependencies, including will not be upgraded unless the version
constraints require it. Ansible is specified with an upper bound on the
version, so typically this will be satisfied during an upgrade and will not
trigger an upgrade.

This change uses the 'latest' state for the pip module to ensure packages are
updated.

Change-Id: Ica38a1cdfb57c4be81468607800b26fdf3209fe7
Story: 2003878
Task: 26737