Move group_vars for switches group to all group
There are various playbook group variables used to configure switches,
defined in files in ansible/group_vars/switches/. In Ansible 2.3
it was possible to override these via inventory group variables.
In Ansible 2.4+ this is no longer possible, and inventory group
variables have a lower precedence than inventory group variables.

To resolve this, this change moves the switch configuration defaults to
the all group, as is done for other global defaults. The slight downside
is that these variables enter the namespace of every host, but there
aren't too many of them so it shouldn't be a problem.

Change-Id: I1a580a7b88a3e088aa7d5129290427320401f46b
Story: 2004278
Task: 27827
(cherry picked from commit 959123681e9ce225d20e9b7ee719af1614739017)