Update grafana-conf role
Ansible Galaxy switched to semantic versioning. In order to get it
to import the latest release, a tag following this scheme was added
after a minor update to the grafana-conf role. This appears to have
broken the accessibility of the original version. This commit moves
to the later release to work around the issue.

Change-Id: I5a0ba9e18bc5db54f899b545b88c319c7922bf9e