Fix overcloud provisioning with newer Bifrost
Recent changes in Bifrost require the following fixes.

Define instance_info as an empty dict in hostvars. Bifrost checks if
this variable is defined and is not an empty dict before using it, but
somehow its value is null if not provided, which results in the followig
error:

    When setting an instance to present, instance_info is a required variable.

Define ironic_api_url instead of letting Bifrost construct it from
hostvars which fail to be gathered in our environment.

Change-Id: I560e0a23097702ff13a9c8ba587f9ebb0ae11c5b